@@ -2,10 +2,10 @@
 
     <PropertyGroup>
         <OutputType>Library</OutputType>
-        <TargetFramework>net6.0</TargetFramework>
+        <TargetFramework>net8.0</TargetFramework>
         <AssemblyName>PostSharp.Engineering.BuildTools</AssemblyName>
         <GenerateDocumentationFile>false</GenerateDocumentationFile>
-        <LangVersion>latest</LangVersion>
+        <LangVersion>preview</LangVersion>
         <Nullable>enable</Nullable>
         <IsPackable>True</IsPackable>
         <!--  SA0001: XML comment analysis is disabled due to project configuration -->
@@ -14,7 +14,7 @@
 
     <ItemGroup>
         <PackageReference Include="AWSSDK.S3" Version="3.7.7.23" />
-        <PackageReference Include="Azure.Identity" Version="1.10.3" />
+        <PackageReference Include="Azure.Identity" Version="1.11.3" />
         <PackageReference Include="Azure.Security.KeyVault.Secrets" Version="4.5.0" />
         <PackageReference Include="JetBrains.Annotations" Version="2022.3.1" />
         <PackageReference Include="Microsoft.Build" Version="$(MSBuildPackagesVersion)" ExcludeAssets="runtime" /> <!-- Runtime assets excluded in favor of Microsoft.Build.Locator package. -->
