@@ -1,7 +1,6 @@
 // Copyright (c) SharpCrafters s.r.o. See the LICENSE.md file in the root directory of this repository root for details.
 
 using Spectre.Console.Cli;
-using System;
 using System.ComponentModel;
 
 namespace PostSharp.Engineering.BuildTools.Dependencies;
@@ -13,7 +12,7 @@ public class ResetDependenciesCommandSettings : ConfigureDependenciesCommandSett
 {
     [Description( "The list of dependencies (given by name by or position in the dependency list) to configure" )]
     [CommandArgument( 1, "[dependencies]" )]
-    public string[] Dependencies { get; protected set; } = Array.Empty<string>();
+    public string[] Dependencies { get; protected set; } = [];
 
     [Description( "Specifies that all dependencies must be configured" )]
     [CommandOption( "--all" )]
