@@ -7,6 +7,7 @@ using PostSharp.Engineering.BuildTools.CodeStyle;
 using PostSharp.Engineering.BuildTools.ContinuousIntegration;
 using PostSharp.Engineering.BuildTools.Csproj;
 using PostSharp.Engineering.BuildTools.Dependencies;
+using PostSharp.Engineering.BuildTools.Docker;
 using PostSharp.Engineering.BuildTools.Git;
 using PostSharp.Engineering.BuildTools.NuGet;
 using PostSharp.Engineering.BuildTools.Tools;
@@ -69,6 +70,27 @@ namespace PostSharp.Engineering.BuildTools
                                 .WithData( product )
                                 .WithDescription( "Bumps the version of this product" );
                         }
+
+                        root.AddBranch(
+                            "docker",
+                            docker =>
+                            {
+                                docker.AddCommand<DockerPrepareCommand>( "prepare" )
+                                    .WithData( product )
+                                    .WithDescription( "Builds an image ready to run the product build." );
+
+                                docker.AddCommand<DockerBuildCommand>( "build" )
+                                    .WithData( product )
+                                    .WithDescription( "Builds the product inside docker." );
+
+                                docker.AddCommand<DockerTestCommand>( "test" )
+                                    .WithData( product )
+                                    .WithDescription( "Runs the product tests inside docker." );
+
+                                docker.AddCommand<DockerInteractiveCommand>( "interactive" )
+                                    .WithData( product )
+                                    .WithDescription( "Opens an interactive PowerShell session inside the docker container." );
+                            } );
 
                         root.AddBranch(
                             "dependencies",
@@ -174,7 +196,7 @@ namespace PostSharp.Engineering.BuildTools
                                         git.AddCommand<GitBulkRenameCommand>( "rename" )
                                             .WithDescription( "Renames all files and directories recursively preserving GIT history." )
                                             .WithExample( new[] { @"""C:\src\Caravela.Compiler""", @"""Caravela""", @"""Metalama""" } );
-                                        
+
                                         git.AddCommand<DownstreamMergeCommand>( "merge-downstream" )
                                             .WithData( product )
                                             .WithDescription( "Merges the code to the subsequent development branch." );
@@ -187,11 +209,10 @@ namespace PostSharp.Engineering.BuildTools
                                             .WithData( product )
                                             .WithDescription(
                                                 "Sets the branch policies of the development and release branch of the current product version." );
-                                        
+
                                         git.AddCommand<PrintBranchPoliciesCommand>( "print-branch-policies" )
                                             .WithData( product )
-                                            .WithDescription(
-                                                "Prints the branch policies currently set for the repository." );
+                                            .WithDescription( "Prints the branch policies currently set for the repository." );
 
                                         git.AddCommand<SetDefaultBranchCommand>( "set-default-branch" )
                                             .WithData( product )
