@@ -38,14 +38,12 @@ namespace PostSharp.Engineering.BuildTools.Dependencies.Model
         /// </summary>
         /// <param name="context"></param>
         /// <param name="settings"></param>
-        /// <param name="configuration"></param>
         /// <returns></returns>
         private bool TryLoadDefaultDependencies(
             BuildContext context,
-            CommonCommandSettings settings,
-            BuildConfiguration configuration )
+            CommonCommandSettings settings )
         {
-            if ( !VersionFile.TryRead( context, settings, configuration, out var versionFile ) )
+            if ( !VersionFile.TryRead( context, settings, out var versionFile ) )
             {
                 return false;
             }
@@ -68,7 +66,7 @@ namespace PostSharp.Engineering.BuildTools.Dependencies.Model
 
             file = new DependenciesOverrideFile( configurationSpecificVersionFilePath, configuration );
 
-            if ( !file.TryLoadDefaultDependencies( context, settings, configuration ) )
+            if ( !file.TryLoadDefaultDependencies( context, settings ) )
             {
                 file = null;
 
@@ -284,7 +282,7 @@ namespace PostSharp.Engineering.BuildTools.Dependencies.Model
             var project = new XElement( "Project", new XAttribute( "InitialTargets", "VerifyProductDependencies" ) );
             var document = new XDocument( project );
 
-            project.Add( new XComment( $"File generated by PostSharp.Engineering {VersionHelper.EngineeringVersion}." ) );
+            project.Add( new XComment( $"File generated by PostSharp.Engineering {VersionHelper.EngineeringVersion}, method {nameof(DependenciesOverrideFile)}.{nameof(this.TrySave)}." ) );
 
             var requiredFiles = new List<string>();
 
@@ -436,6 +434,7 @@ namespace PostSharp.Engineering.BuildTools.Dependencies.Model
             var properties = new XElement( "PropertyGroup" );
             project.Add( properties );
             properties.Add( new XElement( "PostSharpEngineeringExePath", context.Product.BuildExePath ) );
+            properties.Add( new XElement( "PostSharpEngineeringDataDirectory", PathHelper.GetEngineeringDataDirectory() ) );
             properties.Add( new XElement( "ProductFamily", context.Product.ProductFamily.Name ) );
             properties.Add( new XElement( "ProductFamilyVersion", context.Product.ProductFamily.Version ) );
             properties.Add( new XElement( "BuildDate", $"$({context.Product.ProductNameWithoutDot}BuildDate)" ) );
